requirements.txt
fastapi
uvicorn
httpx
google-cloud-storage
google-cloud-aiplatform
